requirements.txt
# Core ML/CV Libraries
ultralytics==8.3.221
openvino==2024.6.0
opencv-python==4.11.0.86
numpy>=1.24.0,<2.0.0

# Object Detection & Tracking
supervision==0.26.1

# MQTT
paho-mqtt==2.1.0

# Utilities
tqdm>=4.66.0
